Compute response distribution indicators

Calculates response distribution indicators for multi-item scales or matrix questions which have the same number of response options for many questions.

Usage

resp_distributions(x, min_valid_responses = 1)

Arguments

x

A data frame containing survey responses in wide format.

min_valid_responses

numeric between 0 and 1. Defines the share of valid responses a respondent must have to calculate response quality indicators. Default is 1.

Value

Returns a data frame with response quality indicators per respondent. Dimensions:

Rows: Equal to number of rows in x.

Columns: 9, one for each response distribution indictator

Details

The functions assumes that the input data frame is structured in the following way:

The data frame is in wide format, meaning each row represents one respondent, each column represents one variable.

The variables are in same the order as the questions respondents saw while taking the survey.

All responses have integer values.

Missing values are set to NA.

The interpretation of the indicators calculated depends on the whether response data of negatively worded questions are reversed or not:

Do not reverse data of negatively worded questions if you want to assess average response patterns (Dunn et al., 2018).

Reverse data of negatively worded questions if you want to assess whether responses are distributed randomly or not with respect to an assumed latent variable (Marjanovic et al., 2015).

The following response distribution indicators are calculated per respondent:

n_valid: number of within person valid answers

n_na: number of within person missing answers

prop_na: proportion of missing responses within respondents

ii_mean: intra-individual mean

ii_median: intra-individual median

ii_median_abs_dev: intra-individual median absolute deviation

ii_sd: intra-individual standard deviation

ii_var: intra-individual variance

mahal: Mahalanobis distance per respondent.

Intra-individual response variability indicators (ii_sd, ii_var) have been proposed to measure insufficient effort responding (Dunn et al., 2018) and to distinguish between random and conscientious responding (Marjanovic et al, 2015). ii_median_abs_dev is added as a robust alternative.

Intra-individual location indicators can be used to asses the average location of responses on a set of questions (ii_mean, ii_median).

Mahalanobis distance is a outlier detection indicator. It represents the distance of a participants responses from the center of a multivariate normal distribution defined by the data of all respondents.

Reponse distribution measures

References

Dunn, Alexandra M., Eric D. Heggestad, Linda R. Shanock, and Nels Theilgard. 2018. “Intra-Individual Response Variability as an Indicator of Insufficient Effort Responding: Comparison to Other Indicators and Relationships with Individual Differences.” Journal of Business and Psychology 33(1):105–21. doi: 10.1007/s10869-016-9479-0.

Marjanovic, Zdravko, Ronald Holden, Ward Struthers, Robert Cribbie, and Esther Greenglass. 2015. “The Inter-Item Standard Deviation (ISD): An Index That Discriminates between Conscientious and Random Responders.” Personality and Individual Differences 84:79–83. doi: 10.1016/j.paid.2014.08.021.

See also

Author

Matthias Roth, Matthias Bluemke & Clemens Lechner

Examples

# A small test data set with ten respondents
# and responses to three survey questions
# with response scales from 1 to 5.
testdata <- data.frame(
  var_a = c(1,4,3,5,3,2,3,1,3,NA),
  var_b = c(2,5,2,3,4,1,NA,2,NA,NA),
  var_c = c(1,2,3,NA,3,4,4,5,NA,NA))

# Calculate response distribution indicators
resp_distributions(testdata) |>
    round(2)
#>    n_valid n_na prop_na wr_mean wr_sd wr_var wr_median wr_median_abs_dev mahal
#> 1        3    0    0.00    1.33  0.58   0.33         1                 0  2.04
#> 2        3    0    0.00    3.67  1.53   2.33         4                 1  1.60
#> 3        3    0    0.00    2.67  0.58   0.33         3                 0  1.38
#> 4        2    1    0.33      NA    NA     NA        NA                NA    NA
#> 5        3    0    0.00    3.33  0.58   0.33         3                 0  0.97
#> 6        3    0    0.00    2.33  1.53   2.33         2                 1  1.38
#> 7        2    1    0.33      NA    NA     NA        NA                NA    NA
#> 8        3    0    0.00    2.67  2.08   4.33         2                 1  1.88
#> 9        1    2    0.67      NA    NA     NA        NA                NA    NA
#> 10       0    3    1.00      NA    NA     NA        NA                NA    NA

# Include respondents with NA values by decreasing the
# necessary number of valid responses per respondent.

resp_distributions(testdata,
                   min_valid_responses = 0.2) |>
    round(2)
#>    n_valid n_na prop_na wr_mean wr_sd wr_var wr_median wr_median_abs_dev mahal
#> 1        3    0    0.00    1.33  0.58   0.33       1.0               0.0  2.27
#> 2        3    0    0.00    3.67  1.53   2.33       4.0               1.0  1.68
#> 3        3    0    0.00    2.67  0.58   0.33       3.0               0.0  1.05
#> 4        2    1    0.33    4.00  1.41   2.00       4.0               1.0  2.21
#> 5        3    0    0.00    3.33  0.58   0.33       3.0               0.0  1.24
#> 6        3    0    0.00    2.33  1.53   2.33       2.0               1.0  1.29
#> 7        2    1    0.33    3.50  0.71   0.50       3.5               0.5  0.71
#> 8        3    0    0.00    2.67  2.08   4.33       2.0               1.0  2.24
#> 9        1    2    0.67    3.00   NaN    NaN       3.0               0.0  0.24
#> 10       0    3    1.00      NA    NA     NA        NA                NA    NA