Compute response style indicators

Calculates response style indicators for multi-item scales

Usage

resp_styles(x, scale_min, scale_max, min_valid_responses = 1, normalize = TRUE)

Arguments

x

A data frame containing survey responses in wide format.

scale_min

numeric. Minimum of scale provided.

scale_max

numeric. Maximum of scale provided.

min_valid_responses

numeric between 0 and 1. Defines the share of valid responses a respondent must have to calculate response style indicators.

normalize

logical. If true, counts of response style indicators will be divided by the number of non-missing responses per respondent.

Value

Returns a data frame with response quality indicators per respondent.

Details

Response style indicators are ... resp_styles is aimed at multi-item scales which use the response options for many questions.

The functions assumes that the input data frame is structured in the following way:

The data frame is in wide format, meaning each row represents one respondent, each column represents one variable.

The variables are in same the order as the questions respondents saw while taking the survey.

Reverse keyed variables are in their original form. No items were recoded.

All responses have integer values.

Questions have the same number of response options.

Missing values are set to NA.

The following response style indicators are calculated per respondent:

MRS: Middle response style (only if scale has a numeric midpoint)

ARS: Acquiescence response style

DRS: Disacquiescence response style

ERS: Extreme response style

NERS: Non-extreme response style

See also