Compute response distribution indices

Calculates response distribution indicators for multi-item scales or matrix questions which have the same number of response options for many questions.

Usage

resp_distributions(x, min_valid_responses = 1)

Arguments

x

A dataframe containing survey responses in wide format.

min_valid_responses

numeric between 0 and 1. Defines the share of valid responses a respondent must have to calculate response quality indicators. Default is 1.

Details

The functions assumes that the input data frame is structured in the following way:

The data frame is in wide format, meaning each row represents one respondent, each column represents one variable.

The variables are in same the order as the questions respondents saw while taking the survey.

Reverse keyed variables are in their original form. No items were recoded.

All responses have integer values.

Missing values are set to NA.

The following response distribution indicators are calculated per respondent:

n_valid: number of within person valid answers

n_na: number of within person missing answers

prop_na: proportion of missing responses within respondents

ii_mean: intra-individual mean over all items

ii_median: intra-individual median

ii_median_abs_dev: intra-individual median absolute deviation

ii_sd: intra-individual standard deviation

ii_var: intra-individual variance

mahal: Mahalanobis distance per respondent. Represents the distance of observations from the center of a multivariate normal distribution.

Reponse distribution measures @returns Returns a data frame with response quality indicators per respondent. Dimensions:

Rows: Equal to number of rows in x.

Columns: 9, one for each response distribution indictator

See also

Author

Matthias Roth, Matthias Bluemke & Clemens Lechner

Examples

# A small test data set with ten respondents
# and responses to three survey questions
# with response scales from 1 to 5.
testdata <- data.frame(
  var_a = c(1,4,3,5,3,2,3,1,3,NA),
  var_b = c(2,5,2,3,4,1,NA,2,NA,NA),
  var_c = c(1,2,3,NA,3,4,4,5,NA,NA))

# Calculate response distribution indicators
resp_distributions(testdata) |>
    round(2)
#>    n_valid n_na prop_na wr_mean wr_sd wr_var wr_median wr_median_abs_dev mahal
#> 1        3    0    0.00    1.33  0.58   0.33         1                 0  2.04
#> 2        3    0    0.00    3.67  1.53   2.33         4                 1  1.60
#> 3        3    0    0.00    2.67  0.58   0.33         3                 0  1.38
#> 4        2    1    0.33      NA    NA     NA        NA                NA    NA
#> 5        3    0    0.00    3.33  0.58   0.33         3                 0  0.97
#> 6        3    0    0.00    2.33  1.53   2.33         2                 1  1.38
#> 7        2    1    0.33      NA    NA     NA        NA                NA    NA
#> 8        3    0    0.00    2.67  2.08   4.33         2                 1  1.88
#> 9        1    2    0.67      NA    NA     NA        NA                NA    NA
#> 10       0    3    1.00      NA    NA     NA        NA                NA    NA

# Include respondents with NA values by decreasing the
# necessary number of valid responses per respondent.

resp_distributions(testdata,
                   min_valid_responses = 0.2) |>
    round(2)
#>    n_valid n_na prop_na wr_mean wr_sd wr_var wr_median wr_median_abs_dev mahal
#> 1        3    0    0.00    1.33  0.58   0.33       1.0               0.0  2.27
#> 2        3    0    0.00    3.67  1.53   2.33       4.0               1.0  1.68
#> 3        3    0    0.00    2.67  0.58   0.33       3.0               0.0  1.05
#> 4        2    1    0.33    4.00  1.41   2.00       4.0               1.0  2.21
#> 5        3    0    0.00    3.33  0.58   0.33       3.0               0.0  1.24
#> 6        3    0    0.00    2.33  1.53   2.33       2.0               1.0  1.29
#> 7        2    1    0.33    3.50  0.71   0.50       3.5               0.5  0.71
#> 8        3    0    0.00    2.67  2.08   4.33       2.0               1.0  2.24
#> 9        1    2    0.67    3.00   NaN    NaN       3.0               0.0  0.24
#> 10       0    3    1.00      NA    NA     NA        NA                NA    NA